+++ libraries/ghc-prim/changelog.md
@@ -1,3 +1,20 @@
+## next (edit as necessary)
+
+- `magicDict` is now defined in `GHC.Magic.Dict` instead of `GHC.Prim`.
+  It now has the type:
+
+        magicDict :: forall {rr :: RuntimeRep} dt st (r :: TYPE rr). (dt => r) -> st -> r
+
+  `magicDict` can now be used without defining an intermediate data type. For example,
+  the `withTypeable` function from `base`'s `Data.Typeable` module can now be defined as:
+
+        withTypeable :: forall k (a :: k) rep (r :: TYPE rep). ()
+                     => TypeRep a -> (Typeable a => r) -> r
+        withTypeable rep k = magicDict @(Typeable a) k rep
+
+  Note that the explicit type application is required, as the call to
+  `magicDict` would be ambiguous otherwise.
+
 ## 0.8.0 (edit as necessary)
 
 - Change array access primops to use type with size maxing the element size:
@@ -23,7 +40,7 @@
 - Add known-key `cstringLength#` to `GHC.CString`. This is just the
   C function `strlen`, but a built-in rewrite rule allows GHC to
   compute the result at compile time when the argument is known.
-  
+
 - In order to support unicode better the following functions in `GHC.CString`
   gained UTF8 counterparts:
 
@@ -47,7 +64,7 @@
         atomicCasAddrAddr# :: Addr# -> Addr# -> Addr# -> State# s -> (# State# s, Addr# #)
         atomicCasWordAddr# :: Addr# -> Word# -> Word# -> State# s -> (# State# s, Word# #)
 
-- Add an explicit fixity for `(~)` and `(~~)`: 
+- Add an explicit fixity for `(~)` and `(~~)`:
 
         infix 4 ~, ~~